requirements.txt
# Core RAG Framework
langchain==0.1.20
langgraph==0.0.40
langchain-openai==0.1.0
langchain-anthropic==0.1.0
langchain-community==0.0.38

# Vector Database
chromadb==0.4.22
sentence-transformers==2.3.1

# Document Processing
pypdf2==3.0.1
pdfplumber==0.10.3
python-docx==1.1.0
pandas==2.2.0
openpyxl==3.1.2

# Web UI
streamlit==1.31.0

# Network Access
smbprotocol==1.12.0

# Configuration Management
pydantic==2.6.0
pydantic-settings==2.1.0
python-dotenv==1.0.1
pyyaml==6.0.1

# Utilities
loguru==0.7.2
tqdm==4.66.1
tenacity==8.2.3

# LLM Providers
openai==1.12.0
anthropic==0.18.1

# Development & Testing
pytest==7.4.4
pytest-asyncio==0.23.4
black==24.1.0
ruff==0.2.0
